project/backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
python-dotenv==1.0.0
scrapy==2.11.0